+++ src/feature/render-csr.ts
@@ -167,6 +167,7 @@ ${overlayHtml}
           ? "globalThis.__DWEB_HMR_DEBUG__ = globalThis.__DWEB_HMR_DEBUG__ ?? true;"
           : ""
       }
+  globalThis.__DWEB_DEV__ = ${isDevCsr};
   globalThis.__DWEB_ROUTES__ = ${JSON.stringify(clientRoutes)};
   globalThis.__DWEB_ENGINE__ = "${engine}";
   globalThis.__DWEB_CONTAINER_ID__ = "${csrOptions.containerId}";